opt: Move unused column rules to separate file
Unused column filtering involves a cohesive set of rules and helper
functions. Separate them out into their own files.

Release note: None
